Further granularize the implementation of bulk_async
